refactor(core): split note content contracts
